# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/certificates/{id}": {
      "get": {
        "operationId": "CertificatesController_getCertificateById",
        "summary": "Get a certificate by id",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Return the certificate.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Certificate"
                }
              }
            }
          },
          "404": {
            "description": "Throws exception if certificate is not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotFoundException"
                }
              }
            }
          }
        },
        "tags": [
          "Certificates"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      },
      "patch": {
        "operationId": "CertificatesController_updateCertificate",
        "summary": "Update a certificate with id",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCertificateDto"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Return the certificate.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Certificate"
                }
              }
            }
          },
          "404": {
            "description": "Throws exception if certificate/cemetery is not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotFoundException"
                }
              }
            }
          }
        },
        "tags": [
          "Certificates"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      },
      "delete": {
        "operationId": "CertificatesController_removeCertificate",
        "summary": "Remove a certificate with id",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Return the certificate.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Certificate"
                }
              }
            }
          },
          "404": {
            "description": "Throws exception if certificate is not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotFoundException"
                }
              }
            }
          }
        },
        "tags": [
          "Certificates"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      }
    },
    "/api/certificates/slug/{slug}": {
      "get": {
        "operationId": "CertificatesController_getCertificateBySlug",
        "summary": "Get a certificate by slug",
        "parameters": [
          {
            "name": "slug",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Return the certificate."
          },
          "404": {
            "description": "Throws exception if certificate is not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotFoundException"
                }
              }
            }
          }
        },
        "tags": [
          "Certificates"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      }
    },
...
    "/api/files/single": {
      "post": {
        "operationId": "FileController_uploadSingleImage",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateColudnleryDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Files"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      }
    },
    "/api/files/multy": {
      "post": {
        "operationId": "FileController_uploadMultyImage",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateMultyCludnleryDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Files"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      }
    },
    "/api/files/certificate-profile/{certificateId}": {
      "post": {
        "operationId": "FileController_createCertificateProfile",
        "summary": "Create a new file",
        "parameters": [
          {
            "name": "certificateId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateFileDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The file has been successfully created.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/File"
                }
              }
            }
          },
          "404": {
            "description": "Throw exception if the certification is not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotFoundException"
                }
              }
            }
          }
        },
        "tags": [
          "Files"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      }
    },
    "/api/files/blog-cntent/{blogContentId}": {
      "post": {
        "operationId": "FileController_createBlogContentImage",
        "summary": "Create a new file",
        "parameters": [
          {
            "name": "blogContentId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateFileDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The file has been successfully created.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/File"
                }
              }
            }
          },
          "404": {
            "description": "Throw exception if the blog content is not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotFoundException"
                }
              }
            }
          }
        },
        "tags": [
          "Files"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      }
    },
    "/api/files/certificate-files/{certificateId}": {
      "patch": {
        "operationId": "FileController_createCertificateFiles",
        "summary": "Create new files",
        "parameters": [
          {
            "name": "certificateId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateFileBodyDto"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "The files has been successfully created/removed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/File"
                }
              }
            }
          },
          "404": {
            "description": "Throw exception if the certification is not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotFoundException"
                }
              }
            }
          }
        },
        "tags": [
          "Files"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      }
    },
    "/api/files/public/{publicId}": {
      "get": {
        "operationId": "FileController_getFileByPublicId",
        "summary": "Get a file by public id",
        "parameters": [
          {
            "name": "publicId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Return the file.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/File"
                }
              }
            }
          },
          "404": {
            "description": "Throw exception if the file is not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotFoundException"
                }
              }
            }
          }
        },
        "tags": [
          "Files"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      }
    },
    "/api/files/{id}": {
      "get": {
        "operationId": "FileController_getFileById",
        "summary": "Get a file by id",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Return the file.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/File"
                }
              }
            }
          },
          "404": {
            "description": "Throw exception if the file is not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotFoundException"
                }
              }
            }
          }
        },
        "tags": [
          "Files"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      },
      "delete": {
        "operationId": "FileController_removeImage",
        "summary": "Get a files by id",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Return the files.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/File"
                }
              }
            }
          },
          "404": {
            "description": "Throw exception if the file is not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotFoundException"
                }
              }
            }
          }
        },
        "tags": [
          "Files"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      }
    },
    "/api/files/certificates/{certificateId}": {
      "get": {
        "operationId": "FileController_getFilesByCertificateId",
        "summary": "Get a files by certificate id",
        "parameters": [
          {
            "name": "certificateId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Return the files.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/File"
                }
              }
            }
          },
          "404": {
            "description": "Throw exception if the file is not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotFoundException"
                }
              }
            }
          }
        },
        "tags": [
          "Files"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      }
    },
    "/api/files/certificates/slug/{slug}": {
      "get": {
        "operationId": "FileController_getFilesBySlug",
        "summary": "Get a files by slug",
        "parameters": [
          {
            "name": "slug",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Return the files.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/File"
                }
              }
            }
          },
          "404": {
            "description": "Throw exception if the file is not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotFoundException"
                }
              }
            }
          }
        },
        "tags": [
          "Files"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "Pomen API",
    "description": "Our API provides a seamless way to create and manage memorial profiles. It allows developers to integrate features for storing and sharing memories, uploading photos and videos, and locating memorial sites. With built-in search functionality, users can easily find profiles by name and explore their biographies, family connections, and important life events. The API supports personalization options, including profile updates and interaction with guestbooks. Designed for scalability and ease of use, our API empowers developers to build meaningful experiences around preserving memories and honoring loved ones.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "File": {
        "type": "object",
        "properties": {}
      },
...
      "User": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "number"
          },
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "profileImageId": {
            "type": "number"
          },
          "profileImage": {
            "$ref": "#/components/schemas/File"
          },
          "gender": {
            "type": "string"
          },
          "dateOfBirth": {
            "format": "date-time",
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          },
          "deletedAt": {
            "format": "date-time",
            "type": "string"
          },
          "lastLoggedInAt": {
            "format": "date-time",
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "phoneNumber": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "isEmailConfirmed": {
            "type": "boolean"
          },
          "roles": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "permissions": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "orders": {
            "$ref": "#/components/schemas/Order"
          },
          "certificates": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Certificate"
            }
          }
        },
        "required": [
          "userId",
          "firstName",
          "lastName",
          "profileImageId",
          "profileImage",
          "gender",
          "dateOfBirth",
          "createdAt",
          "updatedAt",
          "deletedAt",
          "lastLoggedInAt",
          "email",
          "phoneNumber",
          "password",
          "isEmailConfirmed",
          "roles",
          "permissions",
          "orders",
          "certificates"
        ]
      },
      "Gethering": {
        "type": "object",
        "properties": {
          "getheringId": {
            "type": "number"
          },
          "getheringDate": {
            "format": "date-time",
            "type": "string"
          },
          "hour": {
            "type": "number"
          },
          "address": {
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          },
          "deletedAt": {
            "format": "date-time",
            "type": "string"
          },
          "certificateId": {
            "type": "number"
          },
          "certificate": {
            "$ref": "#/components/schemas/Certificate"
          }
        },
        "required": [
          "getheringId",
          "getheringDate",
          "hour",
          "address",
          "createdAt",
          "updatedAt",
          "deletedAt",
          "certificateId",
          "certificate"
        ]
      },
      "Tribute": {
        "type": "object",
        "properties": {
          "tributeId": {
            "type": "number"
          },
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "status": {
            "type": "object"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          },
          "deletedAt": {
            "format": "date-time",
            "type": "string"
          },
          "certificateId": {
            "type": "number"
          },
          "certificate": {
            "$ref": "#/components/schemas/Certificate"
          }
        },
        "required": [
          "tributeId",
          "firstName",
          "lastName",
          "description",
          "email",
          "status",
          "createdAt",
          "updatedAt",
          "deletedAt",
          "certificateId",
          "certificate"
        ]
      },
      "Qrcode": {
        "type": "object",
        "properties": {
          "qrcodeId": {
            "type": "number"
          },
          "value": {
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          },
          "deletedAt": {
            "format": "date-time",
            "type": "string"
          },
          "certificateId": {
            "type": "number"
          },
          "certificate": {
            "$ref": "#/components/schemas/Certificate"
          }
        },
        "required": [
          "qrcodeId",
          "value",
          "createdAt",
          "updatedAt",
          "deletedAt",
          "certificateId",
          "certificate"
        ]
      },
      "Certificate": {
        "type": "object",
        "properties": {
          "certificateId": {
            "type": "number"
          },
          "slug": {
            "type": "string"
          },
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "dateOfBirth": {
            "format": "date-time",
            "type": "string"
          },
          "dateOfDeath": {
            "format": "date-time",
            "type": "string"
          },
          "placeOfBirth": {
            "type": "string"
          },
          "placeOfDeath": {
            "type": "string"
          },
          "timeOfDeath": {
            "type": "number"
          },
          "biography": {
            "type": "string"
          },
          "location": {
            "type": "object"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          },
          "deletedAt": {
            "format": "date-time",
            "type": "string"
          },
          "cemeteryId": {
            "type": "number"
          },
          "cemetery": {
            "$ref": "#/components/schemas/Cemetery"
          },
          "userId": {
            "type": "number"
          },
          "user": {
            "$ref": "#/components/schemas/User"
          },
          "getherings": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Gethering"
            }
          },
          "tributes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Tribute"
            }
          },
          "qrcode": {
            "$ref": "#/components/schemas/Qrcode"
          },
          "orders": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Order"
            }
          },
          "certificateProfileId": {
            "type": "number"
          },
          "profileImage": {
            "$ref": "#/components/schemas/File"
          },
          "files": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/File"
            }
          }
        },
        "required": [
          "certificateId",
          "slug",
          "firstName",
          "lastName",
          "dateOfBirth",
          "dateOfDeath",
          "placeOfBirth",
          "placeOfDeath",
          "timeOfDeath",
          "biography",
          "location",
          "createdAt",
          "updatedAt",
          "deletedAt",
          "cemeteryId",
          "cemetery",
          "userId",
          "user",
          "getherings",
          "tributes",
          "qrcode",
          "orders",
          "certificateProfileId",
          "profileImage",
          "files"
        ]
      },
      "Cemetery": {
        "type": "object",
        "properties": {
          "cemeteryId": {
            "type": "number"
          },
          "address": {
            "type": "string"
          },
          "slug": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "location": {
            "type": "object"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          },
          "deletedAt": {
            "format": "date-time",
            "type": "string"
          },
          "cityId": {
            "type": "number"
          },
          "city": {
            "$ref": "#/components/schemas/City"
          },
          "certificates": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Certificate"
            }
          }
        },
        "required": [
          "cemeteryId",
          "address",
          "slug",
          "name",
          "location",
          "createdAt",
          "updatedAt",
          "deletedAt",
          "cityId",
          "city",
          "certificates"
        ]
      },
...
      "NotFoundException": {
        "type": "object",
        "properties": {}
      },
...
      "UpdateCertificateDto": {
        "type": "object",
        "properties": {}
      },
...
      "FileTypeEnum": {
        "type": "string",
        "description": "Type must be one of the following: image, videos, files",
        "enum": [
          "image",
          "video",
          "document"
        ]
      },
      "CreateColudnleryDto": {
        "type": "object",
        "properties": {
          "file": {
            "type": "string"
          },
          "type": {
            "$ref": "#/components/schemas/FileTypeEnum"
          }
        },
        "required": [
          "file",
          "type"
        ]
      },
      "CreateMultyCludnleryDto": {
        "type": "object",
        "properties": {
          "type": {
            "$ref": "#/components/schemas/FileTypeEnum"
          }
        },
        "required": [
          "type"
        ]
      },
      "CreateFileDto": {
        "type": "object",
        "properties": {
          "height": {
            "type": "number"
          },
          "width": {
            "type": "number"
          },
          "url": {
            "type": "string"
          },
          "publicId": {
            "type": "string"
          },
          "fileExtension": {
            "type": "string"
          },
          "type": {
            "$ref": "#/components/schemas/FileTypeEnum"
          }
        },
        "required": [
          "height",
          "width",
          "url",
          "publicId",
          "fileExtension",
          "type"
        ]
      },
      "CreateFileBodyDto": {
        "type": "object",
        "properties": {
          "filesToAdd": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CreateFileDto"
            }
          },
          "filesToRemove": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "filesToAdd",
          "filesToRemove"
        ]
      },
...
    }
  }
}